+++ src/lib/pages/interact/index.tsx
@@ -7,35 +7,83 @@ import {
   Text,
   Box,
 } from "@chakra-ui/react";
-import { useCallback, useState } from "react";
+import { useRouter } from "next/router";
+import { useCallback, useEffect, useState } from "react";
 
+import { useInternalNavigate } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
 import { LabelText } from "lib/components/LabelText";
 import { ModuleSourceCode } from "lib/components/module";
 import PageContainer from "lib/components/PageContainer";
 import type { IndexedModule } from "lib/services/move/moduleService";
-import { useVerifyModule } from "lib/services/move/moduleService";
-import type { ExposedFunction, HexAddr, Option } from "lib/types";
+import {
+  useAccountModules,
+  useVerifyModule,
+} from "lib/services/move/moduleService";
+import type {
+  ExposedFunction,
+  HexAddr,
+  MoveAccountAddr,
+  Option,
+} from "lib/types";
+import { getFirstQueryParam } from "lib/utils";
 
 import {
   ModuleSelectDrawerTrigger,
   ModuleSelectDrawer,
   FunctionSelectPanel,
   FunctionSelectBody,
+  InteractionTabs,
 } from "./component";
 
-export const Interaction = () => {
+export const Interact = () => {
+  const router = useRouter();
+  const navigate = useInternalNavigate();
   const { isOpen, onClose, onOpen } = useDisclosure();
 
   const [module, setModule] = useState<IndexedModule>();
+  const [selectedType, setSelectedType] = useState<InteractionTabs>(
+    InteractionTabs.VIEW_MODULE
+  );
   const [selectedFn, setSelectedFn] = useState<ExposedFunction>();
+
+  const handleSetSelectedType = (type: string) =>
+    setSelectedType(
+      type === "execute"
+        ? InteractionTabs.EXECUTE_MODULE
+        : InteractionTabs.VIEW_MODULE
+    );
 
   const handleModuleSelect = useCallback(
     (selectedModule: IndexedModule, fn?: ExposedFunction) => {
       setModule(selectedModule);
       setSelectedFn(fn);
+      navigate({
+        pathname: "/interact",
+        query: {
+          address: selectedModule.address,
+          moduleName: selectedModule.moduleName,
+          ...(fn && { functionName: fn.name }),
+        },
+        options: { shallow: true },
+      });
     },
-    []
+    [navigate]
+  );
+
+  const handleFunctionSelect = useCallback(
+    (fn: ExposedFunction) => {
+      setSelectedFn(fn);
+      navigate({
+        pathname: "/interact",
+        query: {
+          ...router.query,
+          functionName: fn.name,
+        },
+        options: { shallow: true },
+      });
+    },
+    [navigate, router.query]
   );
 
   const { data: verificationData } = useVerifyModule({
@@ -43,12 +91,50 @@ export const Interaction = () => {
     moduleName: module?.moduleName,
   });
 
+  const addressParam = getFirstQueryParam(router.query.address);
+  const moduleNameParam = getFirstQueryParam(router.query.moduleName);
+  const functionNameParam = getFirstQueryParam(router.query.functionName);
+  const functionTypeParam = getFirstQueryParam(router.query.functionType);
+  const { refetch } = useAccountModules({
+    address: addressParam as MoveAccountAddr,
+    moduleName: moduleNameParam,
+    functionName: functionNameParam,
+    options: {
+      refetchOnWindowFocus: false,
+      enabled: false,
+      retry: false,
+      onSuccess: (data) => {
+        if (!Array.isArray(data)) {
+          setModule(data);
+          if (functionNameParam) {
+            const fn = [...data.viewFunctions, ...data.executeFunctions].find(
+              (exposedFn) => exposedFn.name === functionNameParam
+            );
+            if (fn) {
+              handleSetSelectedType(fn.is_view ? "view" : "execute");
+              setSelectedFn(fn);
+            }
+          } else if (functionTypeParam)
+            handleSetSelectedType(functionTypeParam);
+        }
+      },
+    },
+  });
+
+  useEffect(() => {
+    if (!!addressParam && !!moduleNameParam) refetch();
+    else {
+      setModule(undefined);
+      setSelectedType(InteractionTabs.VIEW_MODULE);
+      setSelectedFn(undefined);
+    }
+  }, [addressParam, moduleNameParam, refetch]);
+
   return (
     <>
       <PageContainer
         overflow="hidden"
         minH="unset"
-        maxH="calc(100vh - 129px)"
         display="grid"
         gridTemplateRows="auto auto 1fr"
       >
@@ -81,9 +167,7 @@ export const Interaction = () => {
                 </LabelText>
                 <LabelText label="Friends" labelWeight={600}>
                   <Text variant="body2" color="gray.400">
-                    {module?.parsedAbi.friends
-                      ? JSON.stringify(module?.parsedAbi.friends)
-                      : "N/A"}
+                    {JSON.stringify(module.parsedAbi.friends)}
                   </Text>
                 </LabelText>
               </Flex>
@@ -117,6 +201,7 @@ export const Interaction = () => {
           <ModuleSelectDrawer
             isOpen={isOpen}
             onClose={onClose}
+            hexAddress={module?.address}
             handleModuleSelect={handleModuleSelect}
           />
         </Flex>
@@ -128,8 +213,10 @@ export const Interaction = () => {
           {/* Left side */}
           <FunctionSelectPanel
             module={module}
+            tab={selectedType}
+            setTab={setSelectedType}
             selectedFn={selectedFn}
-            setSelectedFn={setSelectedFn}
+            setSelectedFn={handleFunctionSelect}
           />
           {/* Right side */}
           <FunctionSelectBody