+++ src/lib/utils/cw2.ts
@@ -1,7 +1,7 @@
-import type { Option } from "lib/types";
+import type { Nullable, Option } from "lib/types";
 
 export const getCw2Info = (
-  cw2Contract: Option<string | null>,
-  cw2Version: Option<string | null>
+  cw2Contract: Option<Nullable<string>>,
+  cw2Version: Option<Nullable<string>>
 ): Option<string> =>
   cw2Contract && cw2Version ? `${cw2Contract} (${cw2Version})` : undefined;