+++ src/lib/utils/option.ts
@@ -1,4 +1,4 @@
-import type { Option } from "lib/types";
+import type { Nullable, Option } from "lib/types";
 
 export const unwrap = <T>(data: Option<T>, errorMsg?: string): T => {
   if (data === undefined) {
@@ -7,7 +7,10 @@ export const unwrap = <T>(data: Option<T>, errorMsg?: string): T => {
   return data;
 };
 
-export const unwrapAll = <T>(data: Option<T | null>, errorMsg?: string): T => {
+export const unwrapAll = <T>(
+  data: Option<Nullable<T>>,
+  errorMsg?: string
+): T => {
   if (data === undefined || data === null) {
     throw new Error(errorMsg || "Cannot unwrap the given data");
   }